+++ clojure-mode.el
@@ -807,6 +807,18 @@ any number of matches of `clojure--sym-forbidden-rest-chars'.")
                 "\\(\\sw+\\)?")
        (1 font-lock-keyword-face)
        (2 'clojure-type-definition-face nil t))
+      ;; Known other type definition e.g. Schema (s/defschema FooBar ...)
+      (,(concat "(\\(?:" clojure--sym-regexp "/\\)?\\("
+                (regexp-opt '("defschema"))
+                ;; Function declarations
+                "\\)\\>"
+                ;; Any whitespace
+                "[ \r\n\t]*"
+                ;; Possibly type or metadata
+                "\\(?:#?^\\(?:{[^}]*}\\|\\sw+\\)[ \r\n\t]*\\)*"
+                (concat "\\(" clojure--sym-regexp "\\)?"))
+       (1 font-lock-keyword-face)
+       (2 'clojure-type-definition-face nil t))
       ;; Definition with a keyword in the name e.g. spec (s/def :foobar ...)
       (,(concat "(\\(?:" clojure--sym-regexp "/\\)?"
                 "\\(def[^ \r\n\t]*\\)"